feat: Refactor RequesterJob and WorkerJob structs
This commit refactors the RequesterJob and WorkerJob structs in the codebase. The RequesterJob struct now has additional methods for checking if the job is done and waiting for the job to complete. The WorkerJob struct has been modified to remove the commented out code and to add a new method for retrieving the job output. These changes improve the clarity and functionality of the job-related structs.